+++ CHANGELOG.md
@@ -18,8 +18,9 @@ Introduced save state file version 10.
 - App: Avoid crash if the `<profile>/roms/cart` folder is deleted while the emulator is running and the user loads a game that needs to load a cartridge from that folder.
 - CD Block: Properly initialize internal filesystem state and remove unnecessarily strict save state check. Fixes crashes related to the rewind buffer.
 - GUI: Limit maximum size of various windows.
-- Media: Tracks now include the unit sizes along with sector sizes, only needed for CHDs.
 - Media: Allow loading CHDs that don't contain raw sector data, such as those created from ISOs.
+- Media: Restrict ISO loader to files with the .iso extension to prevent users from loading .bin files instead of the .cue files.
+- Media: Tracks now include the unit sizes along with sector sizes, only needed for CHDs.
 - Rewind: Allow varying the size of the state struct. Fixes occasional crashes when rewinding.
 - VDP1: Adjust Y coordinate framebuffer offsets for erase process based on TVMR.TVM. Fixes erase glitches in Grandia when using transparent meshes.
 - VDP2: Clear framebuffer when switching resolutions. Fixes single-frame artifacts in multiple games that switch modes without clearing the screen.